restructure code and move go code to backend folder
