+++ backend/scrapper/utils/datatype.go
