+++ backend/scrapper/utils/functions.go
